requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3

streamlit==1.39.0

chromadb==0.5.5
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2

prometheus-client==0.21.0

black==24.8.0
ruff==0.6.9
pytest==8.3.3
sentence-transformers==3.2.1
